start work on x3232_mooside; gitignore bin/exe files w/o extension
it compiles and it can read/write from/to ttyS0.  Of course this start
is silly... I'm not going to use termios because this program is for
use by me in a specific situation.  See Readme.  As I understand it,
termios etc is for writing a program that can be used in many kinds
of setups.  Also, since I've a MAX3232CPE chip in there, and the chip
effectively makes me use only two lines, TxD and RxD, i.e., without
all the other lines in a RS-232 setup, what would I use termios for?