use termios to setup things
previously, i used putty before i run x3232_mooside. and the simple
and direct read/write on ttyS seems to be working just fine, w/o
need for termios. but, putty must have setup ttyS somehow. so, when,
after a reboot, not running putty first, i run my program, and things
are not working... so, turns out i still need termios to set things up.
then, it's learning time... man termios... and looking for examples and
tutorials on the web... and here's the first version, with termios.

it compiles. and it runs. but sleeping time now, not yet testing with
RPi... have to wait for the morning...